Add exception checking for corrupted files
Some (very very few) of the files in the new agora data appear to be
corrupted. This might be my fault for a reason which shall remain
secret, and I should try re-extracting and re-naming them sometime to
see if that fixes it. For now, the scripts now skip over corrupted
files.